services/deepseek-ocr-mlx/requirements.txt
# MLX framework for Apple Silicon
mlx>=0.20.0
mlx-vlm>=0.1.0

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Image processing
Pillow>=10.2.0

# Data validation
pydantic>=2.6.0

# PDF processing (for batch PDF OCR)
pdf2image>=1.17.0

# HTTP client for testing
httpx>=0.26.0
